[examples] Extract MultiData Module/Loader to common
[ghstack-poisoned]
