+++ examples/common/dataset_utils/__init__.py
@@ -4,7 +4,4 @@
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
 
-from .transforms import *  # noqa F401
-from .utils import *  # noqa F401
-from .imagenet_zeroshot_data import *  # noqa F401
-from .datamodules import *  # noqa F401
+from .multidata import *  # noqa F401
