+++ examples/flava/data/utils.py
@@ -12,7 +12,7 @@ from typing import List
 import requests
 from datasets import concatenate_datasets, load_dataset
 from datasets.utils.file_utils import get_datasets_user_agent
-from definitions import HFDatasetInfo
+from flava.definitions import HFDatasetInfo
 from PIL import Image, UnidentifiedImageError
