+++ examples/flava/utils.py
@@ -4,7 +4,11 @@
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
 
-from definitions import FLAVAArguments, TrainingArguments, TrainingSingleDatasetInfo
+from flava.definitions import (
+    FLAVAArguments,
+    TrainingArguments,
+    TrainingSingleDatasetInfo,
+)
 from hydra.utils import instantiate
 from omegaconf import OmegaConf
